Move validators and types out of WebhookModel.
